Game-Controlled Cameras tutorial
https://docs.unrealengine.com/en-US/Programming/Tutorials/AutoCamera/index.html

Attach a Camera to a moving Actor to create a crane or dolly shot.

Use a single Array  variable to store your cameras, instead of CameraOne and CameraTwo, so you can go through a sequence of any number of cameras, instead of just two.

Instead of using Actor  pointers to store your cameras, create a Structure  that holds the pointer, as well as time before changing the view, and blend time to the new view.