Player-Controlled Cameras: tutorial
https://docs.unrealengine.com/en-US/Programming/Tutorials/PlayerCamera/index.html

Give the player a run key to hold down that will increase the Pawn's movement speed.

Experiment with different ways to mix automatic and input-driven camera movement. This is a very deep area of game development with a lot of room to explore!

Increase, decrease, or remove the lag from your Spring Component to get a better understanding of how much lag can affect your camera's overall feel.

Implement a small amount of periodic motion, possibly slightly randomized or using a Curve asset, to create a handheld feel to your camera.

Give your Camera some amount of automatic turning, so that the camera gradually gets behind the moving player object and faces the direction the player is moving.